# contracts/financial-platform.openapi.json
{
  "components": {
    "schemas": {
      "AdminAssistantProfile": {
        "properties": {
          "api_key_hint": {
            "default": "",
            "title": "Api Key Hint",
            "type": "string"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "connection_id": {
            "default": "",
            "title": "Connection Id",
            "type": "string"
          },
          "model": {
            "default": "",
            "title": "Model",
            "type": "string"
          },
          "provider_name": {
            "default": "",
            "title": "Provider Name",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Updated At"
          }
        },
        "required": [
          "configured"
        ],
        "title": "AdminAssistantProfile",
        "type": "object"
      },
      "AdminAssistantProfileWrite": {
        "properties": {
          "connection_id": {
            "maxLength": 36,
            "minLength": 1,
            "title": "Connection Id",
            "type": "string"
          },
          "model": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "connection_id",
          "model"
        ],
        "title": "AdminAssistantProfileWrite",
        "type": "object"
      },
...
      "AgentModelRequest": {
        "additionalProperties": false,
        "description": "Pi 的 OpenAI Chat Completions 请求白名单。",
        "properties": {
          "connection_id": {
            "anyOf": [
              {
                "maxLength": 36,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Id"
          },
          "max_completion_tokens": {
            "anyOf": [
              {
                "maximum": 32768.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Completion Tokens"
          },
          "max_tokens": {
            "anyOf": [
              {
                "maximum": 32768.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "messages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          },
          "model": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "parallel_tool_calls": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parallel Tool Calls"
          },
          "reasoning_effort": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Effort"
          },
          "response_format": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Format"
          },
          "stream": {
            "default": true,
            "title": "Stream",
            "type": "boolean"
          },
          "stream_options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Options"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "tool_choice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Choice"
          },
          "tools": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 32,
            "title": "Tools",
            "type": "array"
          },
          "top_p": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top P"
          }
        },
        "required": [
          "model",
          "messages"
        ],
        "title": "AgentModelRequest",
        "type": "object"
      },
...
      "AssistantConversationRead": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/AssistantMessageRead"
            },
            "title": "Messages",
            "type": "array"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Updated At"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "AssistantConversationRead",
        "type": "object"
      },
      "AssistantMessageRead": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant",
              "system"
            ],
            "title": "Role",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "session_id",
          "role",
          "content",
          "created_at"
        ],
        "title": "AssistantMessageRead",
        "type": "object"
      },
      "AssistantMessageWrite": {
        "properties": {
          "content": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "role": {
            "enum": [
              "user",
              "assistant",
              "system"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "AssistantMessageWrite",
        "type": "object"
      },
      "AssistantPrepareRequest": {
        "properties": {
          "file_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "File Ids",
            "type": "array"
          },
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "AssistantPrepareRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TaskDraft": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/SkillSummary"
            },
            "title": "Candidates",
            "type": "array"
          },
          "clarification": {
            "default": "",
            "title": "Clarification",
            "type": "string"
          },
          "confidence": {
            "default": 0,
            "title": "Confidence",
            "type": "number"
          },
          "confirmation_text": {
            "default": "",
            "title": "Confirmation Text",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Expires At"
          },
          "file_hashes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "File Hashes",
            "type": "object"
          },
          "files": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              ]
            },
            "title": "Files",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "missing_inputs": {
            "items": {
              "type": "string"
            },
            "title": "Missing Inputs",
            "type": "array"
          },
          "owner_id": {
            "title": "Owner Id",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "requires_approval": {
            "default": false,
            "title": "Requires Approval",
            "type": "boolean"
          },
          "requires_confirmation": {
            "default": true,
            "title": "Requires Confirmation",
            "type": "boolean"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Run Id"
          },
          "skill_id": {
            "title": "Skill Id",
            "type": "string"
          },
          "skill_name": {
            "title": "Skill Name",
            "type": "string"
          },
          "skill_version": {
            "title": "Skill Version",
            "type": "string"
          },
          "source": {
            "default": "manual",
            "enum": [
              "manual",
              "assistant"
            ],
            "title": "Source",
            "type": "string"
          },
          "state": {
            "default": "draft",
            "enum": [
              "draft",
              "ready",
              "expired",
              "consumed"
            ],
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "validation_warnings": {
            "items": {
              "type": "string"
            },
            "title": "Validation Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "owner_id",
          "skill_id",
          "skill_name",
          "skill_version",
          "created_at",
          "updated_at"
        ],
        "title": "TaskDraft",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "财务 Skill 平台受控 API。OpenAPI 是 Next.js 前端类型的唯一来源。",
    "title": "财务 Skill 运行平台",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/assistant-profile": {
      "delete": {
        "operationId": "delete_admin_profile_api_admin_assistant_profile_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Delete Admin Profile",
        "tags": [
          "assistant"
        ]
      },
      "get": {
        "operationId": "get_admin_profile_api_admin_assistant_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAssistantProfile"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Admin Profile",
        "tags": [
          "assistant"
        ]
      },
      "put": {
        "operationId": "put_admin_profile_api_admin_assistant_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminAssistantProfileWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAssistantProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Admin Profile",
        "tags": [
          "assistant"
        ]
      }
    },
...
    "/api/assistant/conversations/{session_id}": {
      "get": {
        "operationId": "assistant_conversation_api_assistant_conversations__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantConversationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assistant Conversation",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/assistant/conversations/{session_id}/messages": {
      "post": {
        "operationId": "append_assistant_conversation_message_api_assistant_conversations__session_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantMessageWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantMessageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Assistant Conversation Message",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/assistant/model": {
      "post": {
        "description": "给服务端 Pi Runtime 提供受部门模型配置约束的 SSE 上游。",
        "operationId": "stream_agent_model_api_assistant_model_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "模型流式响应。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Agent Model",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/assistant/prepare": {
      "post": {
        "operationId": "prepare_api_assistant_prepare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantPrepareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prepare",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/assistant/prepare-from-recommendation": {
      "post": {
        "description": "校验 Pi 工具返回的推荐，并复用现有草稿安全边界。",
        "operationId": "prepare_from_agent_recommendation_api_assistant_prepare_from_recommendation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentPrepareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prepare From Agent Recommendation",
        "tags": [
          "assistant"
        ]
      }
    },
...
    "/api/task-drafts/{draft_id}": {
      "delete": {
        "operationId": "delete_draft_api_task_drafts__draft_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Draft",
        "tags": [
          "assistant"
        ]
      },
      "get": {
        "operationId": "get_draft_api_task_drafts__draft_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Draft",
        "tags": [
          "assistant"
        ]
      },
      "patch": {
        "operationId": "update_draft_api_task_drafts__draft_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskDraftUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Draft",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/task-drafts/{draft_id}/confirm": {
      "post": {
        "operationId": "confirm_draft_api_task_drafts__draft_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Draft",
        "tags": [
          "assistant"
        ]
      }
    },
...
  }
}